Replace brownShoeX.png with brown1ShoeX.png
Note that there is the inclusion of the digit '1' after the word
"brown". This is to match the conventions of the subdirectory structure
used by the modularcharacters assets. The code currently uses partial
name of the image to determine which subfolder the image resides in.

For e.g., brownShoe1.png would be in /Brown/brownShoe1.png whereas
brown1Shoe1.png would be in /Brown 1/brown1Shoe1.png.

This was done to match the existing structure of the assets provided.