Add part menu groups json files
This enables the grouping of character part names (and assets)
that should be categorized together. For e.g., the colors of the
skin_head, skin_neck, skin_arm_left, skin_arm_right, etc. should
really be the same color (usually), and this schema is meant to
help capture this abstract concept for later processing.